requirements-deploy.txt
# Minimal requirements for deployment
# Only essential packages needed to run the application

# Web framework
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.20

# Image processing
opencv-python==4.11.0.86
pillow==11.3.0
numpy==1.26.4

# ML model (YOLO for tab detection)
ultralytics==8.3.214

# Core dependencies
pydantic==2.12.2
starlette==0.48.0

